[rust] rearrange rust support, move D1 specific part to sun20iw1 module
