+++ rust/macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "syterkit-macros"
-version = "0.1.0"
+version = "0.0.0"
 edition = "2021"
 
 [lib]
