+++ rust/src/soc.rs
@@ -1,1 +1,3 @@
+//! Allwinner System-on-Chip (SoC) support.
+
 pub mod sun20iw1;
